requirements.txt
mcp>=0.2.0
mcp[cli]
uv>=0.6.5
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
pandas>=2.0.0
Flask>=2.2.3
SQLAlchemy>=2.0.0
requests>=2.28.2